+++ locales/components/CallForm.yml
@@ -1,0 +1,9 @@
+en:
+  sending: "..."
+  call: Call
+  privacy_html: >
+    Your number will only be used for this call and will never be shared
+    with third parties.
+    <a href="https://www.fightforthefuture.org/privacy/" target="_blank">
+      Privacy Policy</a>
+  error: Sorry, that didn't work. Please check your info and try again.
