+++ locales/components/CallFormModal.yml
@@ -4,11 +4,3 @@ en:
   instructions: >
     We'll provide you with a suggestion of what to say and connect you directly
     with your lawmaker's office.
-  sending: "..."
-  call: Call
-  privacy_html: >
-    Your number will only be used for this call and will never be shared
-    with third parties.
-    <a href="https://www.fightforthefuture.org/privacy/" target="_blank">
-      Privacy Policy</a>
-  error: Sorry, that didn't work. Please check your info and try again.
